requirements.txt
beautifulsoup4==4.11.2
requests==2.31.0
nltk==3.8.1
numpy==1.23.5
pandas==1.5.3
seaborn==0.12.2
matplotlib==3.7.1
wordcloud==1.9.2
spacy==3.7.2
textblob==0.17.1
scikit-learn==1.3.0
streamlit==1.29.0
flask==2.0.1
torch==2.0.1
transformers==4.30.2
python-docx==0.8.11
